Add scan time requirement (#439)
* Add scan time requirement

Let's us set the requirement remaining time to complete a scan
Any scan with a `remaining_time` lower than this requirement raises `ExpiredException`

* Add scan_time_requirement to worker

Uses scan_time_requirement to expire jobs that are not possible.

* Sync duration and timeout

We have a few second buffer on the producer side to account for the time it takes to submit to polyswarmd, and the time it took to get the message from the websocket. 

So, we have been telling the worker they have the full duration to scan, while only listening for duration - timeout. 

This tells the worker it only has as long as we are waiting to scan the file.

* Update grammar

* Whitespace

* Show it takes seconds.

* Use updated message

* Typo